fix(video): enable Bink playback on Linux

+++ video/videoservices.cpp
@@ -1173,8 +1173,8 @@ bool CVideoCommonServices::CalculateVideoDimensions( int videoWidth, int videoHe
 				}
 				else	// can scale variably..
 				{
-					float Xfactor = ( displayWidth / curWidth );
-					float Yfactor = ( displayHeight / curHeight );
+					float Xfactor = ( (float)displayWidth / curWidth );
+					float Yfactor = ( (float)displayHeight / curHeight );
 					float scale = MIN( Xfactor, Yfactor );
 					
 					curWidth = (int)  ( curWidth * scale + 0.35f );
